second commit

+++ modules/task_1/pyatckin_n_avg_of_vector/avg_of_vector.h
@@ -3,11 +3,10 @@
 #define MODULES_TASK_1_PYATCKIN_N_AVG_OF_VECTOR_AVG_OF_VECTOR_H_
 
 #include <vector>
-#include <string>
 
-std::vector<double> getRandomVector(int  sz);
-double getParallelOperations(std::vector<double>& vec,
-	int size_vec);
-double getSequentialOperations(std::vector<double>& vec, int size_vec);
+std::vector<int> getRandomVector(int  sz);
+int getParallelOperations(std::vector<int>& vec,
+    int size_vec);
+int getSequentialOperations(std::vector<int>& vec, int size_vec);
 
 #endif  // MODULES_TASK_1_PYATCKIN_N_AVG_OF_VECTOR_AVG_OF_VECTOR_H_
